pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100
profile = "black"

[tool.poetry]
name = "fastapi-starter"
version = "0.0.3"
description = "Extensible FastAPI application to provide consistent authentication."
authors = ["William Sawyer <william@situ.com.au>"]
readme = "README.md"
repository = "https://github.com/SituDevelopment/fastapi-starter"
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.10.0"

bcrypt = "^4.1.2"
fastapi_another_jwt_auth = "^0.1.10"
fief-client = { version = "^0.18.6", extras = ["fastapi"] }
mangum = "^0.17.0"
pydantic = { version = "^2.6.3", extras = ["email"] }
PyMySQL = "^1.1.0"
python-multipart = "^0.0.9"
requests = "^2.31.0"
SQLAlchemy = "^2.0.25"

[tool.poetry.group.dev.dependencies]

black = "*"
isort = "*"

[tool.poetry.group.test.dependencies]

pylint = "*"
pytest = "*"

[tool.pylint.design]
allowed-redefined-builtins = ["id"]
# FastAPI dependencies require a generic resource ID parameter

[tool.pylint.main]
exclude-too-few-public-methods = ["Base", "BaseModel"]
# Pydantic schemas and SQLAlchemy models are not designed to have public methods
